+++ src/main/java/com/zunf/tankbattletcpserver/model/bo/TankBO.java
@@ -42,7 +42,7 @@ public class TankBO implements Serializable {
     /**
      * 速度 px/tick
      */
-    private Integer speed;
+    private Double speed;
 
     /**
      * 生命值
